+++ src/fixtures/modules/Some_Test_Svelte.svelte
@@ -1,1 +1,9 @@
+<script lang="ts" context="module">
+	export const a = 'ok';
+</script>
+
+<script lang="ts">
+	// ...
+</script>
+
 Some_Test_Svelte.svelte contents