+++ src/fixtures/modules/some_test_server.ts
@@ -1,0 +1,5 @@
+import {Some_Test_Svelte_Ts} from './some_test_svelte_ts.svelte.js';
+
+export const some_test_server = 'some_test_server';
+
+export const Rexported_Some_Test_Svelte_Ts = Some_Test_Svelte_Ts;