+++ src/fixtures/modules/some_test_svelte_js.svelte.js
@@ -1,0 +1,3 @@
+export class Some_Test_Svelte_Js {
+	a = $state('ok');
+}