+++ src/lib/esbuild_plugin_external_worker.ts
@@ -2,7 +2,7 @@ import * as esbuild from 'esbuild';
 import type {Logger} from '@ryanatkn/belt/log.js';
 import {basename} from 'node:path';
 import {cwd} from 'node:process';
-import type {CompileOptions, PreprocessorGroup} from 'svelte/compiler';
+import type {CompileOptions, PreprocessorGroup, ModuleCompileOptions} from 'svelte/compiler';
 
 import {print_build_result, to_define_import_meta_env} from './esbuild_helpers.js';
 import {resolve_specifier} from './resolve_specifier.js';
@@ -18,6 +18,7 @@ export interface Options {
 	build_options: esbuild.BuildOptions;
 	dir?: string;
 	svelte_compile_options?: CompileOptions;
+	svelte_compile_module_options?: ModuleCompileOptions;
 	svelte_preprocessors?: PreprocessorGroup | PreprocessorGroup[];
 	alias?: Record<string, string>;
 	base_url?: Parsed_Sveltekit_Config['base_url'];
@@ -36,6 +37,7 @@ export const esbuild_plugin_external_worker = ({
 	dir = cwd(),
 	svelte_compile_options,
 	svelte_preprocessors,
+	svelte_compile_module_options,
 	alias,
 	base_url,
 	assets_url,
@@ -64,7 +66,12 @@ export const esbuild_plugin_external_worker = ({
 						ambient_env,
 					}),
 					esbuild_plugin_sveltekit_shim_alias({dir, alias}),
-					esbuild_plugin_svelte({dir, svelte_compile_options, svelte_preprocessors}),
+					esbuild_plugin_svelte({
+						dir,
+						svelte_compile_options,
+						svelte_compile_module_options,
+						svelte_preprocessors,
+					}),
 					esbuild_plugin_sveltekit_local_imports(),
 				],
 				define: to_define_import_meta_env(dev, base_url),