+++ src/lib/gro_plugin_server.ts
@@ -140,8 +140,17 @@ export const gro_plugin_server = ({
 				private_prefix,
 				public_prefix,
 				svelte_compile_options,
+				svelte_compile_module_options,
 				svelte_preprocessors,
 			} = await init_sveltekit_config(sveltekit_config ?? dir);
+
+			// TODO hacky
+			if (svelte_compile_options.generate === undefined) {
+				svelte_compile_options.generate = 'server';
+			}
+			if (svelte_compile_module_options.generate === undefined) {
+				svelte_compile_module_options.generate = 'server';
+			}
 
 			const {outbase, outdir, outname} = outpaths(dev);
 
@@ -178,6 +187,7 @@ export const gro_plugin_server = ({
 						build_options,
 						dir,
 						svelte_compile_options,
+						svelte_compile_module_options,
 						svelte_preprocessors,
 						alias,
 						base_url,
@@ -188,7 +198,12 @@ export const gro_plugin_server = ({
 						ambient_env,
 						log,
 					}),
-					esbuild_plugin_svelte({dir, svelte_compile_options, svelte_preprocessors}),
+					esbuild_plugin_svelte({
+						dir,
+						svelte_compile_options,
+						svelte_compile_module_options,
+						svelte_preprocessors,
+					}),
 					esbuild_plugin_sveltekit_local_imports(),
 				],
 				define: to_define_import_meta_env(dev, base_url),