+++ src/lib/sveltekit_config.ts
@@ -1,5 +1,5 @@
 import type {Config as SveltekitConfig} from '@sveltejs/kit';
-import type {CompileOptions, PreprocessorGroup} from 'svelte/compiler';
+import type {CompileOptions, PreprocessorGroup, ModuleCompileOptions} from 'svelte/compiler';
 import {join} from 'node:path';
 import {cwd} from 'node:process';
 
@@ -42,7 +42,8 @@ export interface Parsed_Sveltekit_Config {
 	env_dir: string | undefined;
 	private_prefix: string | undefined;
 	public_prefix: string | undefined;
-	svelte_compile_options: CompileOptions | undefined;
+	svelte_compile_options: CompileOptions;
+	svelte_compile_module_options: ModuleCompileOptions;
 	svelte_preprocessors: PreprocessorGroup | PreprocessorGroup[] | undefined;
 }
 
@@ -73,7 +74,8 @@ export const init_sveltekit_config = async (
 	const private_prefix = kit?.env?.privatePrefix;
 	const public_prefix = kit?.env?.publicPrefix;
 
-	const svelte_compile_options = sveltekit_config?.compilerOptions;
+	const svelte_compile_options = sveltekit_config?.compilerOptions ?? {};
+	const svelte_compile_module_options = {}; // TODO from `kit`? or subset of `svelte_compile_options`?
 	const svelte_preprocessors = sveltekit_config?.preprocess;
 
 	return {
@@ -88,6 +90,7 @@ export const init_sveltekit_config = async (
 		private_prefix,
 		public_prefix,
 		svelte_compile_options,
+		svelte_compile_module_options,
 		svelte_preprocessors,
 	};
 };